refactor(services): extract web skill governance, evidence and registry clients
Sixteen methods across three contexts that hold no shared mutable state: skill
tool inspection and drift, skill evolution evidence, and agent registry reads.

Each was already its own connected component in the state map, so the cut rule is
satisfied without any bridging. `purgedEvidenceScopes` moves into the evidence
module; `resetWebEvidenceForTest` is re-exported so its test seam keeps working.

`normalizeWebPath` and `normalizeWebSkillLocation` move to `web-skill-location.ts`
because the composition root's skill CRUD and the extracted drift client must
derive the same canonical form from one function, not two copies.

`syncSkillDrift`, `getSkillEvolutionSeedLineage` and `purgeSkillEvolutionEvidence`
each call a sibling through `this`, so they declare `this: AgentService`. That
also keeps `getSkillOverview` — which stays in the root and calls
`this.detectSkillDrift` — resolving through the composed object.
